+++ composer.json
@@ -32,7 +32,7 @@
   },
   "autoload": {
     "psr-4": {
-      "WooCommerceSerialNumbers\\": ["src/","lib/"]
+      "WooCommerceSerialNumbers\\": ["includes/","src/","lib/"]
     }
   },
   "config": {
